can filter works outside of Emb. Sharepoint, need to interface with can1_recv somehow?
