+++ Tests/Inc/can1_recv_entries.h
@@ -1,4 +1,6 @@
+// CAN_RECV_ENTRY(ID, SIZE, CIRCULAR)
 
-CAN_RECV_ENTRY(0x001, 5)
-CAN_RECV_ENTRY(0x002, 8)
-CAN_RECV_ENTRY(0x003, 3)
+CAN_RECV_ENTRY(0x001, 5, false)
+CAN_RECV_ENTRY(0x002, 8, false)
+CAN_RECV_ENTRY(0x003, 3, false)
+CAN_RECV_ENTRY(0x004, 4, true)
